move from CompressorBuilder -> Compressor, bump rust
