+++ benches/compress.rs
@@ -10,7 +10,7 @@ use std::{
     path::Path,
 };
 
-use criterion::{criterion_group, criterion_main, Criterion, Throughput};
+use criterion::{Criterion, Throughput, criterion_group, criterion_main};
 
 use curl::easy::Easy;
 use fsst::Compressor;
