+++ rust-toolchain.toml
@@ -1,4 +1,4 @@
 [toolchain]
-channel = "nightly-2024-08-14"
+channel = "nightly-2025-02-24"
 components = ["rust-src", "rustfmt", "clippy"]
 profile = "minimal"
