+++ src/builder.rs
@@ -5,8 +5,8 @@
 //! [FSST Paper]: https://www.vldb.org/pvldb/vol13/p2649-boncz.pdf
 
 use crate::{
-    advance_8byte_word, compare_masked, lossy_pht::LossyPHT, Code, Compressor, Symbol,
-    FSST_CODE_BASE, FSST_CODE_MASK,
+    Code, Compressor, FSST_CODE_BASE, FSST_CODE_MASK, Symbol, advance_8byte_word, compare_masked,
+    lossy_pht::LossyPHT,
 };
 use std::cmp::Ordering;
 use std::collections::BinaryHeap;
@@ -72,7 +72,7 @@ struct CodesIterator<'a> {
     reference: usize,
 }
 
-impl<'a> Iterator for CodesIterator<'a> {
+impl Iterator for CodesIterator<'_> {
     type Item = u16;
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -298,104 +298,7 @@ impl Default for CompressorBuilder {
     }
 }
 
-fn validate_symbol_order(symbol_lens: &[u8]) {
-    // Ensure that the symbol table is ordered by length, 23456781
-    let mut expected = 2;
-    for (idx, &len) in symbol_lens.iter().enumerate() {
-        if expected == 1 {
-            assert_eq!(
-                len, 1,
-                "symbol code={idx} should be one byte, was {len} bytes"
-            );
-        } else {
-            if len == 1 {
-                expected = 1;
-            }
-
-            // we're in the non-zero portion.
-            assert!(
-                len >= expected,
-                "symbol code={idx} breaks violates FSST symbol table ordering"
-            );
-            expected = len;
-        }
-    }
-}
-
 impl CompressorBuilder {
-    /// Rebuild a compressor from an existing symbol table.
-    ///
-    /// This will not attempt to optimize or re-order the codes.
-    pub fn rebuild_from(
-        symbols: impl AsRef<[Symbol]>,
-        symbol_lens: impl AsRef<[u8]>,
-    ) -> Compressor {
-        let symbols = symbols.as_ref();
-        let symbol_lens = symbol_lens.as_ref();
-
-        assert_eq!(
-            symbols.len(),
-            symbol_lens.len(),
-            "symbols and lengths differ"
-        );
-        assert!(symbols.len() <= 254, "symbol table len must be <= 254");
-        validate_symbol_order(symbol_lens);
-
-        // Insert the symbols in their given order into the FSST lookup structures.
-        let symbols = symbols.to_vec();
-        let lengths = symbol_lens.to_vec();
-        let mut lossy_pht = LossyPHT::new();
-
-        // Initialize the codes_two_byte table.
-        let mut codes_two_byte = Vec::with_capacity(65_535);
-        for tb in 0..=(255 * 255) {
-            codes_two_byte.push(Code::new_escape(tb as u8));
-        }
-
-        // Insert all of the one-byte symbols first.
-        for (code, (&symbol, &len)) in symbols.iter().zip(lengths.iter()).enumerate() {
-            if len == 1 {
-                codes_two_byte[symbol.first_byte() as usize] = Code::new_symbol(code as u8, 1);
-            }
-        }
-
-        // Insert all of the two-byte symbols over the one-byte
-        for (code, (&symbol, &len)) in symbols.iter().zip(lengths.iter()).enumerate() {
-            if len == 2 {
-                codes_two_byte[symbol.first2() as usize] = Code::new_symbol(code as u8, 2);
-            }
-        }
-
-        // Insert the 3+ byte symbols into the Lossy PHT
-        for (code, (&symbol, &len)) in symbols.iter().zip(lengths.iter()).enumerate() {
-            if len >= 3 {
-                lossy_pht.insert(symbol, len as usize, code as u8);
-            }
-        }
-
-        // Find the position of the first code that has a suffix later in the table
-        let mut has_suffix_code = symbols.len() as u8;
-        for (code, (&symbol, &len)) in symbols.iter().zip(lengths.iter()).enumerate() {
-            if len != 2 {
-                break;
-            }
-            let rest = &symbols[code..];
-            if rest.iter().any(|&other| symbol.first2() == other.first2()) {
-                has_suffix_code = code as u8;
-                break;
-            }
-        }
-
-        Compressor {
-            n_symbols: symbols.len() as u8,
-            symbols,
-            lengths,
-            codes_two_byte,
-            lossy_pht,
-            has_suffix_code,
-        }
-    }
-
     /// Attempt to insert a new symbol at the end of the table.
     ///
     /// # Panics
@@ -949,8 +852,7 @@ impl Ord for Candidate {
 
 #[cfg(test)]
 mod test {
-    use crate::{builder::CodesBitmap, Compressor, CompressorBuilder, Symbol, ESCAPE_CODE};
-    use std::{iter, mem};
+    use crate::{Compressor, ESCAPE_CODE, builder::CodesBitmap};
 
     #[test]
     fn test_builder() {
@@ -1027,58 +929,5 @@ mod test {
     fn test_bitmap_invalid() {
         let mut map = CodesBitmap::default();
         map.set(512);
-    }
-
-    #[test]
-    fn test_symbols_good() {
-        let symbols_u64: &[u64] = &[
-            24931, 25698, 25442, 25699, 25186, 25444, 24932, 25188, 25185, 25441, 25697, 25700,
-            24929, 24930, 25443, 25187, 6513249, 6512995, 6578786, 6513761, 6513507, 6382434,
-            6579042, 6512994, 6447460, 6447969, 6382178, 6579041, 6512993, 6448226, 6513250,
-            6579297, 6513506, 6447459, 6513764, 6447458, 6578529, 6382180, 6513762, 6447714,
-            6579299, 6513508, 6382436, 6513763, 6578532, 6381924, 6448228, 6579300, 6381921,
-            6382690, 6382179, 6447713, 6447972, 6513505, 6447457, 6382692, 6513252, 6578785,
-            6578787, 6578531, 6448225, 6382177, 6382433, 6578530, 6448227, 6381922, 6578788,
-            6579044, 6382691, 6512996, 6579043, 6579298, 6447970, 6447716, 6447971, 6381923,
-            6447715, 97, 98, 100, 99, 97, 98, 99, 100,
-        ];
-        let symbols: &[Symbol] = unsafe { mem::transmute(symbols_u64) };
-        let lens: Vec<u8> = iter::repeat_n(2u8, 16)
-            .chain(iter::repeat_n(3u8, 61))
-            .chain(iter::repeat_n(1u8, 8))
-            .collect();
-
-        let compressor = CompressorBuilder::rebuild_from(symbols, lens);
-        let built_symbols: &[u64] = unsafe { mem::transmute(compressor.symbol_table()) };
-        assert_eq!(built_symbols, symbols_u64);
-    }
-
-    #[should_panic]
-    #[test]
-    fn test_symbols_bad() {
-        let symbols: &[u64] = &[
-            24931, 25698, 25442, 25699, 25186, 25444, 24932, 25188, 25185, 25441, 25697, 25700,
-            24929, 24930, 25443, 25187, 6513249, 6512995, 6578786, 6513761, 6513507, 6382434,
-            6579042, 6512994, 6447460, 6447969, 6382178, 6579041, 6512993, 6448226, 6513250,
-            6579297, 6513506, 6447459, 6513764, 6447458, 6578529, 6382180, 6513762, 6447714,
-            6579299, 6513508, 6382436, 6513763, 6578532, 6381924, 6448228, 6579300, 6381921,
-            6382690, 6382179, 6447713, 6447972, 6513505, 6447457, 6382692, 6513252, 6578785,
-            6578787, 6578531, 6448225, 6382177, 6382433, 6578530, 6448227, 6381922, 6578788,
-            6579044, 6382691, 6512996, 6579043, 6579298, 6447970, 6447716, 6447971, 6381923,
-            6447715, 97, 98, 100, 99, 97, 98, 99, 100,
-        ];
-        let lens: Vec<u8> = iter::repeat_n(2u8, 16)
-            .chain(iter::repeat_n(3u8, 61))
-            .chain(iter::repeat_n(1u8, 8))
-            .collect();
-
-        let mut builder = CompressorBuilder::new();
-        for (symbol, len) in symbols.iter().zip(lens.iter()) {
-            let symbol = Symbol::from_slice(&symbol.to_le_bytes());
-            builder.insert(symbol, *len as usize);
-        }
-        let compressor = builder.build();
-        let built_symbols: &[u64] = unsafe { std::mem::transmute(compressor.symbol_table()) };
-        assert_eq!(built_symbols, symbols);
     }
 }
